+++ src/main/java/edu/itu/cavabunga/advice/ExceptionHandlerAdvice.java
@@ -1,9 +1,9 @@
-package edu.itu.cavabunga.core.advice;
+package edu.itu.cavabunga.advice;
 
-import edu.itu.cavabunga.core.controller.wrapper.ErrorResponse;
-import edu.itu.cavabunga.core.exception.IcalNotFound;
-import edu.itu.cavabunga.core.exception.ParticipantConflict;
-import edu.itu.cavabunga.core.exception.ParticipantNotFound;
+import edu.itu.cavabunga.controller.wrapper.ErrorResponse;
+import edu.itu.cavabunga.exception.IcalNotFound;
+import edu.itu.cavabunga.exception.ParticipantConflict;
+import edu.itu.cavabunga.exception.ParticipantNotFound;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
