+++ src/main/java/edu/itu/cavabunga/exception/AuthenticationNotFound.java
@@ -1,4 +1,4 @@
-package edu.itu.cavabunga.core.exception;
+package edu.itu.cavabunga.exception;
 
 public class AuthenticationNotFound extends RuntimeException{
     public AuthenticationNotFound(){
