+++ src/main/java/edu/itu/cavabunga/exception/IcalNotFound.java
@@ -1,4 +1,4 @@
-package edu.itu.cavabunga.core.exception;
+package edu.itu.cavabunga.exception;
 
 public class IcalNotFound extends RuntimeException{
     public IcalNotFound(){
