+++ src/main/java/edu/itu/cavabunga/exception/ParticipantConflict.java
@@ -1,4 +1,4 @@
-package edu.itu.cavabunga.core.exception;
+package edu.itu.cavabunga.exception;
 
 public class ParticipantConflict extends RuntimeException {
     public ParticipantConflict(){
